# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Saatvik Jain Aahar Gruh - Tiffin Service API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /snacks:
    get:
      operationId: listSnacks
      tags: [snacks]
      summary: List active snacks
      parameters:
        - name: snack_category_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Snacks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Snack"
  /snacks/{id}:
    get:
      operationId: getSnack
      tags: [snacks]
      summary: Get a snack by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Snack detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snack"
  /admin/snacks:
    get:
      operationId: adminListSnacks
      tags: [admin]
      summary: Admin - list all snacks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All snacks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Snack"
    post:
      operationId: createSnack
      tags: [admin]
      summary: Admin - create snack
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnackBody"
      responses:
        "201":
          description: Snack created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snack"
  /admin/snacks/{id}:
    put:
      operationId: updateSnack
      tags: [admin]
      summary: Admin - update snack
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnackBody"
      responses:
        "200":
          description: Snack updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snack"
    delete:
      operationId: deleteSnack
      tags: [admin]
      summary: Admin - soft delete snack
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/snacks/{id}/toggle:
    patch:
      operationId: toggleSnackStatus
      tags: [admin]
      summary: Admin - toggle snack active/inactive
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snack"

  # INVENTORY
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # AUTH
...
    SnackCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
        status:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sortOrder
        - status
...
    Snack:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: string
        imageUrl:
          type: string
        snackCategoryId:
          type: integer
        weight:
          type: string
        status:
          type: integer
        createdAt:
          type: string
          format: date-time
        SnackCategory:
          $ref: '#/components/schemas/SnackCategory'
        Inventory:
          $ref: '#/components/schemas/Inventory'
      required:
        - id
        - name
        - price
    SnackBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        snackCategoryId:
          type: integer
        weight:
          type: string
        status:
          type: integer
      required:
        - name
        - price
        - snackCategoryId
    Inventory:
      type: object
      properties:
        snackId:
          type: integer
        quantity:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required:
        - snackId
        - quantity
...