# lib/api-zod/src/generated/types/snackBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saatvik Jain Aahar Gruh - Tiffin Service API
 * OpenAPI spec version: 0.1.0
 */

export interface SnackBody {
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  snackCategoryId: number;
  weight?: string;
  status?: number;
}